requirements.txt
amqp==5.2.0
anyio==4.3.0
asgiref==3.7.2
billiard==4.2.0
boto3==1.34.62
botocore==1.34.62
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
# Django==5.0.3
django-redis==5.4.0
django-storages==1.14.2
flower==2.0.1
h11==0.14.0
httptools==0.6.1
humanize==4.9.0
idna==3.6
jmespath==1.0.1
kombu==5.3.6
msgpack==1.0.8
mysql==0.0.3
mysqlclient==2.2.4
pillow==10.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.1
spotipy==2.23.0
sqlparse==0.4.4
tornado==6.4
tzdata==2024.1
urllib3==1.26.16
uvicorn==0.28.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.1
Django~=5.0.3
djangorestframework~=3.15.1
automium_web~=0.1.1
pathlib~=1.0.1
django-cors-headers==4.4.0
PyJWT~=2.3.0